Wrap main logic in `try`-`except` block
I just experienced the second case of an exception causing the main
process to fail and me not noticing for a week. I have put the scraping
logic into a dedicated function and put that in a `try` block. I also
added two new logging statements to make behavior easier to evaluate.