Add `<atom:link>` element to provide feed's URL
RSS does not have a way for a feed to store its own URL, but the
"atom:link" section of the RSS Best Practices Profile[^1] recommends
using an element from Atom for that purpose.

This resolves the final recommendation from the W3C Feed Validation
Service.

As part of the implementation, I created a module to handle runtime
configuration because I did not want to hardcode my personal feed URL in
a repo that will be public.

[^1]: https://www.rssboard.org/rss-profile#namespace-elements-atom-link