Reorganize `data` directory
The directory contained only two files, `cache.json` and `feed.rss`.
This is fine when there is only a single shim, but to handle multiple
shims, more structure is needed. There are now separate directories for
cache and feed files, and each shim has its own location within it.

This will require changes to my `docker-compose.yml` file to point to
the new feed directory. The cache files do not be publicly exposed, and
having an extra `feeds/` component in the URL is redundant.