Factor out Comics Kingdom shim into its own path
This commit was much larger than I planned, but this *is* a major
architectural change. I created a shim ABC to define a common interface
 and moved as much logic into it as possible. The remaining logic was
put in the Comics Kingdom shim, and that was genericized to work on
comics other than _Rae the Doe_.

With this commit, I can say that I am almost ready to start implementing
other shims.