Migrate from `setup.py` to `pyproject.toml`
`pyproject.toml` is the latest standard, and using it removes the need
to explicitly import `setuptools` as part of building the package.

I still rely on a `setuptools`-specific extension to pull in the
`requirements.txt` file, but that seems to be the best way to handle
building OCI images for now.

The new TOML code was automatically generated using `pdm import setup.cfg`
and modifying the output to fit my needs. It was much easier than doing
the transfer by hand.