Switch base OCI image from Debian to Alpine
Alpine is significantly smaller than even "slim" Debian images, so
development and deployment should go quicker if I use one.

I was using `3.12.7-slim` (based on Debian bookworm), which was 124 MB.
I am now using `3.12.7-alpine3.20`, which is 46.9 MB.

This drops the size of the final image from 138 MB to 61 MB. It's not
much, but it should still help some. It has the added benefit of having
fewer vulnerabilities.

Changing the base OS required one related build change: I need to run
the server with `sh` because `bash` is not installed by default.